refactor: move filesystem storage to canonical package
