+++ easyuse_anima/infrastructure/filesystem/__init__.py
@@ -1,0 +1,3 @@
+"""Filesystem infrastructure owned by the canonical package."""
+
+__all__ = ()
