+++ tests/test_registry_scanner_safety.py
@@ -47,6 +47,9 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/infrastructure/comfy/capabilities.py",
     "easyuse_anima/infrastructure/comfy/invocation.py",
     "easyuse_anima/infrastructure/comfy/resources.py",
+    "easyuse_anima/infrastructure/filesystem/__init__.py",
+    "easyuse_anima/infrastructure/filesystem/atomic_json.py",
+    "easyuse_anima/infrastructure/filesystem/paths.py",
     "easyuse_anima/lora/__init__.py",
     "easyuse_anima/lora/metadata.py",
     "easyuse_anima/lora/preset.py",
